requirements.txt
customtkinter
cryptography
bcrypt
pillow
image
pyinstaller
sqlalchemy
